Add WASM SIMD128 decode path and fix 32-bit literals header overflow
Add a wasm32 SIMD128 decode path mirroring the existing aarch64 NEON
path (both 128-bit). Core SIMD primitives (wildcopy, copy_match,
common_prefix_len) in crates/core/src/simd/wasm32/simd128.rs; fused
FSE decode+execute loop in crates/decode/src/simd_decode/wasm32/decode.rs.
Dispatch via CpuTier::Wasm32Simd128 (compile-time detection via
cfg!(target_feature = "simd128")), gated by not(feature = "paranoid").

Fix pre-existing 32-bit overflow in parse_compressed_header format-3
branch (literals.rs:78): (data[4] as usize) << 28 overflows on wasm32
where usize is 32 bits, corrupting regen_size/compressed_size and
causing BadHuffmanStream on any block with a 5-byte literals header.
Fixed by accumulating into u64 before truncating to usize.

Includes bench/wasm/ crate for WASM encode+decode benchmarking via
wasmtime, with optional C zstd comparison (--features czstd).

+++ crates/decode/src/literals.rs
@@ -75,12 +75,12 @@ fn parse_compressed_header(data: &[u8]) -> Result<(usize, usize, usize, usize), 
             if data.len() < 5 {
                 return Err(DecompressError::CorruptLiterals);
             }
-            let both = (header_byte as usize >> 4)
-                | ((data[1] as usize) << 4)
-                | ((data[2] as usize) << 12)
-                | ((data[3] as usize) << 20)
-                | ((data[4] as usize) << 28);
-            Ok((both & 0x3FFFF, both >> 18, 4, 5))
+            let both = (header_byte as u64 >> 4)
+                | ((data[1] as u64) << 4)
+                | ((data[2] as u64) << 12)
+                | ((data[3] as u64) << 20)
+                | ((data[4] as u64) << 28);
+            Ok(((both & 0x3FFFF) as usize, (both >> 18) as usize, 4, 5))
         }
         _ => unreachable!(),
     }
